kubectl-datadog: consolidate aws-auth into common/awsauth/
`common/aws/aws-auth.go` was misnamed: it manipulates a Kubernetes
ConfigMap (`kube-system/aws-auth`), not an AWS-SDK resource. Together
with `guess/aws-auth.go` (read-only presence check), it belongs in a
dedicated `common/awsauth/` package: both files operate on the same
ConfigMap.

Function names lose the now-redundant `AwsAuth` prefix:
  - `IsAwsAuthConfigMapPresent` -> `IsConfigMapPresent`
  - `EnsureAwsAuthRole`         -> `EnsureRole`
  - `RemoveAwsAuthRole`         -> `RemoveRole`

`common/aws/` now contains only cloudformation.go (pure AWS-SDK).
